requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
psycopg[binary]==3.2.10
python-dotenv==1.0.1
pydantic==2.9.2
email-validator==2.2.0

# testes
pytest==8.3.2
httpx==0.27.2

